+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 - [GUI] Download errors dialog improvements (#4502 by: HebaruSan)
 - [GUI] Fix symlinks being generated on Linux (#4519 by: HebaruSan)
 - [ConsoleUI] Only scroll ConsoleUI log messages for small popups (#4547 by: HebaruSan)
+- [Multiple] Exclude other versions of mods already in changeset from satisfying relationships (#4550 by: HebaruSan)
 
 ### Internal